DEV-1827 Loosen initial required args checks
The change that was made previously to allow a JSON file containing the
runs (instead of a runs directory) could only be unit tested. However
when run in-app there was a check against the runs directory that was
still breaking the in-GCP clinical loader.

This change loosens the argument checking to allow the SampleLoader
class to do all the checking. It knows about both arguments and will
require one and only one of the runs JSON/runs directory to be given.