DEV-2039 Throw on bad Shallow input, not exit
The main method in CreateShallowSeqDB is now called from the
`shallow-reporter` component to automatically load shallow data when
their pipeline runs complete.

`shallow-reporter` is a long-lived process that waits for successive
runs, so it cannot be exiting every time there is a problem with one of
the runs. It can now handle an unexpected situation properly and
continue running.