DEV-2086 Throw on bad input, not System.exit(1)
The main method in PatientReporterApplication is now called from the new
patient-reporter application to automatically load patient data when
pipeline runs complete.

The GCP patient-reporter is a long-lived process that waits for successive
runs, so it cannot be exiting every time there is a problem with one of the
runs. With this change it can handle an unexpected situation properly and
continue running.

Also add a shading plugin for the patient-reporter JAR to allow us to
build a version of the GCP patient-reporter that behaves reliabily at
runtime.