DEV-2199 Use exact (not major.minor) pipeline ver
This is to accommodate the patient-reporter that runs in GCP as in that
environment there will be a three-part version string rather than the
major.minor we've had when running directly on datastore.

Also backport the shading for the patient-reporter JAR that is required
when including it as a dependency in the GCP wrapped version, and which
is already in the master branch.