DEV-2408 Remove static date from ReportResources
This static field is being used as a global variable for the generation date
of the patient report. It's causing a bug in the GCP-based reporting because
in that situation the process remains running for quite a long time and the
static date drifts from the current date.

Eliminate this global variable in favour of a parameter to the reports that
is injected at the top level. Since it's no longer static the value of this
can be set each time the `.main()` method is called by the controlling
process.

Also in an only tangentially related change I've modified the test for
`CommandOption.kt` as this does not work on Linux, the `cd` command it was
using for a command that should exist does not actually exist as a
standalone executable on Linux. `more` should be available as a binary on
Windows, Linux and Mac.