DEV-2623 Tumor only support in health checker
Refactor the health checker to split the tumor/reference pairs and null checks on tumor, and
group together the per-sample arguments. Loosen restrictions on CLI args to allow for a
null ref as well as a null tumor.

ps. Germline only is already supported as "single sample", so we simply change some logging there.