DEV-2739: Create reporting DB JSON always
Before this change, this JSON was not created
for COLO samples or for samples without a cohort,
since they are probably not actual production samples.
Whether this JSON is printed is only relevant for production
samples, for which we always want the JSON. So we may as well
always create it and simply ignore the JSON when running
non-production samples manually.