DEV-2739: Add -allow_default_cohort_config argument
Previously, when no cohort was configured in LIMS for a sample,
behaviour depended on the tumor sample name format. If the name
started with "COLO", then a hardcoded COLO-config would be used.
Otherwise, an error would be raised. Now, default behaviour is to
always raise an error when no cohort is configured. If instead
the new argument is provided, then a warning is logged
and a hardcoded default config is used,
where the only change from the previous COLO-config is
a change of cohort name from "COLO" to "DEFAULT".
The goal is to reduce reliance on sample name format and to
allow patient-reporter to run on non-COLO, non-LIMS samples.