INFRA-266 Reverse sample exists check to CheckSampleAbsentFromDatabase
The first usage of this class is to check whether a sample does NOT
exist in the database yet, so we don't overwrite data. It seems
therefore more logical to reverse the check so "the sample is absent
from the database" becomes the common expected behavior and "the sample
is not absent from the database" becomes the erroneous behavior that's
logged as a warning.