INFRA-541 Changes to Cuppa invocation
After discussion with Luan I am switching to the existing main class for
the prediction runner, it calls the prep already. I'm also attempting
some improvements to the classes that run the external processes, Luan
says there are some problems with reliability there.

I do not think a separate thread was required and expect that may have
been part of the problem, though do not properly understand everything
yet.  Also the exception handling was a bit confusing and I have
attempted to simplify that just so I can understand what is happening.

Also add R to the Docker image.