INFRA-541 Update Cuppa Dockerfile (again)
Make it look more like the others, which are using `micromamba` to
install the R dependencies.

Also switch back to the directory structure used by Oncoanalyser and
remove the entrypoint, both of which should make it broadly compatible
with that project.

Despite the previous statement I think the location of the Python
component has changed, because it changed in the JAR file, not sure
whether we should port it back to "chart" so leaving it as "pycuppa".