INFRA-541 Remove wget and procps installs
I actually thought these were required for the running of micromamba but
it must have been I made a mistake somewhere. They were definitely used
originally for fetching the JARs but that is not relevant anymore.

Extract the common JDK installation step for the images based on the
Microconda base as that is definitely going to be pretty heavy. Shuffle
up the ordering for the add of the JAR so we don't build a new image
every single time the JAR changes.

Further refinement could be done to the remaining Microconda steps to
have one per line. As they are already listed in alphabetical order this
might result in a few fewer layers being built but I think much less
valuable than the easy change for the JDK package.