INFRA-541 Add deps for circos
I expected these to be pulled in automatically as this is a
Conda-managed dependency but maybe the graphics library is optional.
Seems unlikely as the program does not run without it.

I have had to do this in the past with the OS-level dependencies but am
going to avoid the temptation to add in all the dependencies I had to do
in the past without verifying they are in fact not installed on the
container.

So this is likely not to be the lass commit to do with this issue.