INFRA-541 Base container for circos
Annoying. This is one way to share the Circos resources that are
required to run the charts: we have our own configuration, data files
and fonts which are referenced directly from the files in the repository
for tools that do the charting.

Another way to do this would be with a tarball or recursive ADD
statements in the Dockerfiles but this would result in a multi-stage
build because there is no way to prepare something like that from within
the Dockerfile. Also files to be copied must be done from within the
"context" of the Dockerfile, essentially the same directory. Relative
paths like "../" will result in a build error.

So I don't see this as a terrible solution though it does have some
suboptimal elements.